[Windows] use explicitly 64-bit type

+++ src/metaSMT/DirectSolver_Context.hpp
@@ -63,8 +63,8 @@ namespace metaSMT {
         , Expr1 value
         , Expr2 bw
     ) {
-      const unsigned long val   = proto::value(value);
-      const unsigned long width = proto::value(bw);
+      const uint64_t val   = proto::value(value);
+      const unsigned width = proto::value(bw);
 
       return SolverContext::operator() ( tag,
         boost::any(boost::make_tuple(val, width))
@@ -77,8 +77,8 @@ namespace metaSMT {
         , Expr1 value
         , Expr2 bw
     ) {
-      const long val            = proto::value(value);
-      const unsigned long width = proto::value(bw);
+      const int64_t val    = proto::value(value);
+      const unsigned width = proto::value(bw);
 
       return  SolverContext::operator() ( tag, 
         boost::any(boost::make_tuple(val, width))
